1. vbo prepared 忘记true了 2 implement IBO
I haven't used OpenGL from Java so I'm speculating here,
but there's a good chance that you can just use the negative numbers whose binary reprentation
is the same as the unsigned positive numbers you really want.
You're giving GL some byte pairs and telling it to interpret them as unsigned,
and as long as they have the right value when interpreted that way, it should work.
It doesn't matter if Java thought they meant something different when it stored those bits in memory.
If you're iterating, just ignore the wraparound and keep on incrementing. When you get to -1, you're done.
If you're calculating the index numbers as ints (which don't have this range problem) and then casting to short,
subtract 65536 from any number that's greater than 32767.